Run bin/jaxrs-resteasy-eap-petstore-server.sh
