--- samples/server/petstore/jaxrs-resteasy/eap/src/gen/java/org/openapitools/api/DefaultApi.java
@@ -1,264 +1,0 @@
-package org.openapitools.api;
-
-import org.openapitools.model.*;
-
-import io.swagger.annotations.ApiParam;
-import io.swagger.jaxrs.*;
-
-import java.io.File;
-import java.util.List;
-import java.util.Map;
-import org.openapitools.model.ModelApiResponse;
-import org.openapitools.model.Order;
-import org.openapitools.model.Pet;
-import org.openapitools.model.User;
-
-import java.util.List;
-import java.util.Map;
-
-import java.io.InputStream;
-
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.SecurityContext;
-import javax.ws.rs.*;
-import javax.validation.constraints.*;
-import org.jboss.resteasy.plugins.providers.multipart.MultipartFormDataInput;
-
-@Path("/default")
-
-
-@io.swagger.annotations.Api(description = "the default API")
-
-public interface DefaultApi  {
-   
-    @POST
-    
-    @Consumes({ "application/json", "application/xml" })
-    
-    @io.swagger.annotations.ApiOperation(value = "Add a new pet to the store", notes = "", response = Void.class, authorizations = {
-        @io.swagger.annotations.Authorization(value = "petstore_auth", scopes = {
-            @io.swagger.annotations.AuthorizationScope(scope = "write:pets", description = "modify pets in your account"),
-            @io.swagger.annotations.AuthorizationScope(scope = "read:pets", description = "read your pets")
-        })
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 405, message = "Invalid input", response = Void.class) })
-    public Response addPet(@ApiParam(value = "Pet object that needs to be added to the store" ,required=true) Pet pet,@Context SecurityContext securityContext);
-    @POST
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Create user", notes = "This can only be done by the logged in user.", response = Void.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Void.class) })
-    public Response createUser(@ApiParam(value = "Created user object" ,required=true) User user,@Context SecurityContext securityContext);
-    @POST
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Creates list of users with given input array", notes = "", response = Void.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Void.class) })
-    public Response createUsersWithArrayInput(@ApiParam(value = "List of user object" ,required=true) List<User> user,@Context SecurityContext securityContext);
-    @POST
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Creates list of users with given input array", notes = "", response = Void.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Void.class) })
-    public Response createUsersWithListInput(@ApiParam(value = "List of user object" ,required=true) List<User> user,@Context SecurityContext securityContext);
-    @DELETE
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Delete purchase order by ID", notes = "For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors", response = Void.class, tags={ "store", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid ID supplied", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 404, message = "Order not found", response = Void.class) })
-    public Response deleteOrder( @PathParam("orderId") String orderId,@Context SecurityContext securityContext);
-    @DELETE
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Deletes a pet", notes = "", response = Void.class, authorizations = {
-        @io.swagger.annotations.Authorization(value = "petstore_auth", scopes = {
-            @io.swagger.annotations.AuthorizationScope(scope = "write:pets", description = "modify pets in your account"),
-            @io.swagger.annotations.AuthorizationScope(scope = "read:pets", description = "read your pets")
-        })
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid pet value", response = Void.class) })
-    public Response deletePet( @PathParam("petId") Long petId,@ApiParam(value = "" )@HeaderParam("api_key") String apiKey,@Context SecurityContext securityContext);
-    @DELETE
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Delete user", notes = "This can only be done by the logged in user.", response = Void.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid username supplied", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 404, message = "User not found", response = Void.class) })
-    public Response deleteUser( @PathParam("username") String username,@Context SecurityContext securityContext);
-    @GET
-    
-    
-    @Produces({ "application/xml", "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Finds Pets by status", notes = "Multiple status values can be provided with comma separated strings", response = Pet.class, responseContainer = "List", authorizations = {
-        @io.swagger.annotations.Authorization(value = "petstore_auth", scopes = {
-            @io.swagger.annotations.AuthorizationScope(scope = "write:pets", description = "modify pets in your account"),
-            @io.swagger.annotations.AuthorizationScope(scope = "read:pets", description = "read your pets")
-        })
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Pet.class, responseContainer = "List"),
-        
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid status value", response = Void.class) })
-    public Response findPetsByStatus( @NotNull @QueryParam("status") List<String> status,@Context SecurityContext securityContext);
-    @GET
-    
-    
-    @Produces({ "application/xml", "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Finds Pets by tags", notes = "Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.", response = Pet.class, responseContainer = "List", authorizations = {
-        @io.swagger.annotations.Authorization(value = "petstore_auth", scopes = {
-            @io.swagger.annotations.AuthorizationScope(scope = "write:pets", description = "modify pets in your account"),
-            @io.swagger.annotations.AuthorizationScope(scope = "read:pets", description = "read your pets")
-        })
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Pet.class, responseContainer = "List"),
-        
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid tag value", response = Void.class) })
-    public Response findPetsByTags( @NotNull @QueryParam("tags") List<String> tags,@Context SecurityContext securityContext);
-    @GET
-    
-    
-    @Produces({ "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Returns pet inventories by status", notes = "Returns a map of status codes to quantities", response = Integer.class, responseContainer = "Map", authorizations = {
-        @io.swagger.annotations.Authorization(value = "api_key")
-    }, tags={ "store", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Map.class, responseContainer = "Map") })
-    public Response getInventory(@Context SecurityContext securityContext);
-    @GET
-    
-    
-    @Produces({ "application/xml", "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Find purchase order by ID", notes = "For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions", response = Order.class, tags={ "store", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Order.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid ID supplied", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 404, message = "Order not found", response = Void.class) })
-    public Response getOrderById( @Min(1) @Max(5) @PathParam("orderId") Long orderId,@Context SecurityContext securityContext);
-    @GET
-    
-    
-    @Produces({ "application/xml", "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Find pet by ID", notes = "Returns a single pet", response = Pet.class, authorizations = {
-        @io.swagger.annotations.Authorization(value = "api_key")
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Pet.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid ID supplied", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 404, message = "Pet not found", response = Void.class) })
-    public Response getPetById( @PathParam("petId") Long petId,@Context SecurityContext securityContext);
-    @GET
-    
-    
-    @Produces({ "application/xml", "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Get user by user name", notes = "", response = User.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = User.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid username supplied", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 404, message = "User not found", response = Void.class) })
-    public Response getUserByName( @PathParam("username") String username,@Context SecurityContext securityContext);
-    @GET
-    
-    
-    @Produces({ "application/xml", "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Logs user into the system", notes = "", response = String.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = String.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid username/password supplied", response = Void.class) })
-    public Response loginUser( @NotNull @QueryParam("username") String username, @NotNull @QueryParam("password") String password,@Context SecurityContext securityContext);
-    @GET
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Logs out current logged in user session", notes = "", response = Void.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Void.class) })
-    public Response logoutUser(@Context SecurityContext securityContext);
-    @POST
-    
-    
-    @Produces({ "application/xml", "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "Place an order for a pet", notes = "", response = Order.class, tags={ "store", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = Order.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid Order", response = Void.class) })
-    public Response placeOrder(@ApiParam(value = "order placed for purchasing the pet" ,required=true) Order order,@Context SecurityContext securityContext);
-    @PUT
-    
-    @Consumes({ "application/json", "application/xml" })
-    
-    @io.swagger.annotations.ApiOperation(value = "Update an existing pet", notes = "", response = Void.class, authorizations = {
-        @io.swagger.annotations.Authorization(value = "petstore_auth", scopes = {
-            @io.swagger.annotations.AuthorizationScope(scope = "write:pets", description = "modify pets in your account"),
-            @io.swagger.annotations.AuthorizationScope(scope = "read:pets", description = "read your pets")
-        })
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid ID supplied", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 404, message = "Pet not found", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 405, message = "Validation exception", response = Void.class) })
-    public Response updatePet(@ApiParam(value = "Pet object that needs to be added to the store" ,required=true) Pet pet,@Context SecurityContext securityContext);
-    @POST
-    
-    @Consumes({ "application/x-www-form-urlencoded" })
-    
-    @io.swagger.annotations.ApiOperation(value = "Updates a pet in the store with form data", notes = "", response = Void.class, authorizations = {
-        @io.swagger.annotations.Authorization(value = "petstore_auth", scopes = {
-            @io.swagger.annotations.AuthorizationScope(scope = "write:pets", description = "modify pets in your account"),
-            @io.swagger.annotations.AuthorizationScope(scope = "read:pets", description = "read your pets")
-        })
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 405, message = "Invalid input", response = Void.class) })
-    public Response updatePetWithForm( @PathParam("petId") Long petId,@ApiParam(value = "Updated name of the pet", defaultValue="null")@FormParam("name")  String name,@ApiParam(value = "Updated status of the pet", defaultValue="null")@FormParam("status")  String status,@Context SecurityContext securityContext);
-    @PUT
-    
-    
-    
-    @io.swagger.annotations.ApiOperation(value = "Updated user", notes = "This can only be done by the logged in user.", response = Void.class, tags={ "user", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 400, message = "Invalid user supplied", response = Void.class),
-        
-        @io.swagger.annotations.ApiResponse(code = 404, message = "User not found", response = Void.class) })
-    public Response updateUser( @PathParam("username") String username,@ApiParam(value = "Updated user object" ,required=true) User user,@Context SecurityContext securityContext);
-    @POST
-    
-    @Consumes({ "multipart/form-data" })
-    @Produces({ "application/json" })
-    @io.swagger.annotations.ApiOperation(value = "uploads an image", notes = "", response = ModelApiResponse.class, authorizations = {
-        @io.swagger.annotations.Authorization(value = "petstore_auth", scopes = {
-            @io.swagger.annotations.AuthorizationScope(scope = "write:pets", description = "modify pets in your account"),
-            @io.swagger.annotations.AuthorizationScope(scope = "read:pets", description = "read your pets")
-        })
-    }, tags={ "pet", })
-    @io.swagger.annotations.ApiResponses(value = { 
-        @io.swagger.annotations.ApiResponse(code = 200, message = "successful operation", response = ModelApiResponse.class) })
-    public Response uploadFile(MultipartFormDataInput input, @PathParam("petId") Long petId,@Context SecurityContext securityContext);
-}
