+++ samples/server/petstore/jaxrs-resteasy/eap/src/main/java/org/openapitools/api/RestApplication.java
@@ -6,7 +6,9 @@ import javax.ws.rs.core.Application;
 import java.util.Set;
 import java.util.HashSet;
 
-import org.openapitools.api.impl.DefaultApiServiceImpl;
+import org.openapitools.api.impl.PetApiServiceImpl;
+import org.openapitools.api.impl.StoreApiServiceImpl;
+import org.openapitools.api.impl.UserApiServiceImpl;
 
 @ApplicationPath("/")
 public class RestApplication extends Application {
@@ -14,7 +16,9 @@ public class RestApplication extends Application {
 
     public Set<Class<?>> getClasses() {
         Set<Class<?>> resources = new HashSet<Class<?>>();
-        resources.add(DefaultApiServiceImpl.class);
+        resources.add(PetApiServiceImpl.class);
+        resources.add(StoreApiServiceImpl.class);
+        resources.add(UserApiServiceImpl.class);
 
         return resources;
     }
