+++ samples/server/petstore/jaxrs-resteasy/eap/src/main/java/org/openapitools/api/impl/StoreApiServiceImpl.java
@@ -1,0 +1,35 @@
+package org.openapitools.api.impl;
+
+import org.openapitools.api.*;
+import org.openapitools.model.*;
+
+
+import java.util.Map;
+import org.openapitools.model.Order;
+
+import java.util.List;
+
+import java.io.InputStream;
+
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
+
+
+public class StoreApiServiceImpl implements StoreApi {
+      public Response deleteOrder(String orderId,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response getInventory(SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response getOrderById(Long orderId,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response placeOrder(Order order,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+}
