+++ samples/server/petstore/jaxrs-resteasy/eap/src/main/java/org/openapitools/api/impl/UserApiServiceImpl.java
@@ -1,0 +1,51 @@
+package org.openapitools.api.impl;
+
+import org.openapitools.api.*;
+import org.openapitools.model.*;
+
+
+import java.util.List;
+import org.openapitools.model.User;
+
+import java.util.List;
+
+import java.io.InputStream;
+
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
+
+
+public class UserApiServiceImpl implements UserApi {
+      public Response createUser(User user,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response createUsersWithArrayInput(List<User> user,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response createUsersWithListInput(List<User> user,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response deleteUser(String username,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response getUserByName(String username,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response loginUser(String username,String password,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response logoutUser(SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+      public Response updateUser(String username,User user,SecurityContext securityContext) {
+      // do some magic!
+      return Response.ok().build();
+  }
+}
